mircode/requirements.txt
Django==3.0.8
psycopg2==2.8.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
django-allauth==0.42.0
djangorestframework-jwt==1.11.0